CT-7777
(cherry picked from commit 33deddf9181e781ec6a40aaab1e1eead34e9461e)
(cherry picked from commit 0b32db4b0b5a3fd0a4883995bffa2a06f6420131)
(cherry picked from commit 0f8fe884b457734b812716235ab976bacfbc1225)

+++ i21Database/dbo/Stored Procedures/uspCTImportContractFromCsv.sql
@@ -147,7 +147,7 @@ BEGIN
 	DECLARE @dtmContractDate DATETIME
 	DECLARE @intCommodityId INT
 	DECLARE @intCommodityUOMId INT
-	DECLARE @dblQuantity INT
+	DECLARE @dblQuantity NUMERIC(18, 6)
 	DECLARE @intSalespersonId INT
 	DECLARE @ysnSigned BIT
 	DECLARE @strContractNumber NVARCHAR(200)
